// angular-src/src/models/Role.ts
/*
 * @Description: Model for Role
 * @Author: Guozhi Tang
 * @Date: 2019-07-24 10:50:13
 * @Github: https://github.com/GuozhiTang/Bio-WebApp
 * @LastEditors: Guozhi Tang
 * @LastEditTime: 2019-08-30 14:44:54
 */
export class Role {
  moduleName: String;
  className: String;
  liquid_class: String;
  role: String;
  reagent: Boolean;
  id: String;
}